Moved WCProductProperty* to propertyviews package
