+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/products/viewholders/PropertyViewHolder.kt
@@ -3,8 +3,8 @@ package com.woocommerce.android.ui.products.viewholders
 import android.view.ViewGroup
 import android.widget.LinearLayout
 import com.woocommerce.android.R
-import com.woocommerce.android.ui.products.WCProductPropertyView
 import com.woocommerce.android.ui.products.models.ProductProperty.Property
+import com.woocommerce.android.ui.products.propertyviews.WCProductPropertyView
 
 class PropertyViewHolder(parent: ViewGroup) : ProductPropertyViewHolder(parent, R.layout.product_property_view) {
     fun bind(item: Property) {
