+++ WooCommerce/src/main/res/layout/product_property_button_view.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<com.woocommerce.android.ui.products.WCProductPropertyButtonView
+<com.woocommerce.android.ui.products.propertyviews.WCProductPropertyButtonView
     xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="wrap_content" />
