+++ WooCommerce/src/main/res/layout/product_property_cardview.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<com.woocommerce.android.ui.products.WCProductPropertyCardView
+<com.woocommerce.android.ui.products.propertyviews.WCProductPropertyCardView
     xmlns:android="http://schemas.android.com/apk/res/android"
     style="@style/Woo.Card"
     android:layout_width="match_parent"
