+++ WooCommerce/src/main/res/layout/product_property_read_more_view.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<com.woocommerce.android.ui.products.WCProductPropertyReadMoreView
+<com.woocommerce.android.ui.products.propertyviews.WCProductPropertyReadMoreView
     xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
